Move the pthread code into c.rs so that it gets linked in.
