+++ c-scape/src/lib.rs
@@ -12,7 +12,6 @@ mod data;
 mod environ;
 mod error_str;
 mod exit;
-mod pthread;
 mod unwind;
 
 type GlobalAlloc = wee_alloc::WeeAlloc<'static>;
